Move image check so it is being used beforehand
